python_api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
scipy==1.11.4
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0

